Evaluate Functions Completed. All functions can be evaluated. Backpropagation remains to be created eval code for.
Side effect: If changes to error srings are made, errors will be thrown. This is
somewhat ideal.

We will now create evaluate functions for each new function in main.py.

Some changes have been made by AI. We consider it okay to use AI in evaluate.py,
if the code has been read through and verified to be accurate. However, we will
not use AI in main.py.